move upload button to archived page. trigger refresh upon uploaded.
