+++ src/app/recording-list/archived/archived-recording-list.component.html
@@ -1,6 +1,7 @@
 <div class="card-pf-heading-kebab">
   <div class="pull-right recording-list-toolbar">
     <button class="btn btn-default" title="Refresh" (click)="refreshList()"><span class="fa fa-refresh"></span></button>
+    <button class="btn btn-default" title="Upload" (click)="recordingUpload()"><span class="fa fa-upload"></span></button>
   </div>
 </div>
 <h2 class="card-pf-title">
