+++ src/app/recording-list/current/current-recording-list.component.html
@@ -12,7 +12,6 @@
       </label>
       <button class="btn btn-default" title="Create Recording..." (click)="openRecordingForm()"><span class="fa pficon-add-circle-o"></span></button>
       <button class="btn btn-default" title="Refresh" (click)="refreshList()"><span class="fa fa-refresh"></span></button>
-      <button class="btn btn-default" title="Upload" (click)="uploadRecording()"><span class="fa fa-upload"></span></button>
     </div>
   </div>
   <h2 class="card-pf-title">
